fix: Restrict Kelvin values to steps of 100
The mirek (or mired) to kelvin conversion can suffer from precision
mismatches that can lead to computed Kelvin values slightly above
the expected maximum in certain situations.

The platform convention is to clamp to steps of 100, so we implement
that here when converting from the color temp value we get from the
Hue API in to the Kelvin value we use on the platform.