fix(hue): Inject refresh when transitioning online
We had reports of an issue where a child device that onboards as offline, and then
transitions to online, was in a state where it wouldn't be controllable due to the fact
that no attribute events had been emitted for the device (since it onboarded as offline).

This change injects a refresh capability command event in to the device whenever it is
marked `online` as a result of a connectivity event on the Hue Bridge's SSE stream.

We also clean up a few other places where we should be using injected refreshes instead
of using the handlers directly, though there are a few places where using the handlers
directly is still preferred.