Change method of determining color mode
Set the color mode according to the most recent attribute handler that
runs and check this after each device init rather than reading the
ColorMode from the device.

The original issue is due to subscribing to multiple attributes that can
individually control the color of the device without considering that
there is not a guaranteed order that the attribute handlers will run in.
The changes in this commit address this by having special behavior to
check the color mode following driver init.