Read proper attributes after reading color mode
During init, read the color mode before subscribing. In the color mode
attribute handler, send out read requests for the corresponding
attributes. Then, in the handlers for those attributes, ignore the first
value and wait until the attribute handler runs again before emitting
the colorControl capability. This ensures that we only emit this
capability based on the correct values.