fix: Address bug in color temp range emit logic
The existing logic utilized the device persistent fields to determine
if a range update was needed, as to avoid un-necessary attribute emits.

This would lead to existing devices not ever emitting the min/max color
temp range values if the saved min/max matched what was reported by the
API.

This is adjusted to use the cached latest state for the device instead,
so that we can detect the `nil` if the device has never emitted a value
for its range.