fix(hue): Make rounding logic platform consistent
There was a slight bug with the previous port of the rounding logic
from Matter/Zigbee over to Hue because I was reading two different
places in the code and getting them confused in my head.

This change does the following:

1. Only use the 100 step size constraint when computing the min/max
values instead of everywhere. This is in line with how this is handled
elsewhere on the platform.
2. Use normal rounding everywhere else.