Add support for the color mode attribute (#1772)
In certain situations such as after a driver restart, the hue and saturation attributes of the colorControl capability may be overwritten by device attributes that are not currently controlling the color of the device. This change reads the ColorMode attribute during init to check the current color mode and then reads the appropriate attributes based on the result before emitting the hue and saturation capabilities.