Matter Switch: Support fan/light devices
This change adds support for fan/switch combination devices. A new
multicomponent profile contains the fan capability in the main component
as well as capabilities for a color light in a secondary comoponent. New
logic is also introduced in the matter switch driver to support the
component mapping for this profile.