Add dynamic handling of thermostatOperatingState for Room AC device type
Dynamic handling for the thermostatOperatingState capability was
included by a previous PR for the thermostat and air purifier device
types. This PR extends that change to the Room AC device type, to add
new profiles excluding this capability as well as dynamic profiling to select
one of these profiles if the device does not support the optional
ThermostatRunningState attribute.