Add dynamic handling of thermostatOperatingState for the Room AC device type (#2005)
Dynamic handling for the thermostatOperatingState capability was
included by a previous PR for the thermostat and air purifier device
types. This change extends this to the Room AC device type, to add
a new profile excluding this capability as well as dynamic profiling to select
the profile if the device does not support the optional
ThermostatRunningState attribute.