Only support nostate for existing Room AC profiles (#2009)
Currently, it would be possible for match_profile to choose a profile
that does not currently exist for the Room AC device type. New profiles
could be added and the new logic in this PR could be removed in the
future once the total number of allowed profiles is increased.