Fix wrong colorControl values emitted after device power cycle
Following a device power cycle, hub power cycle, or device
going offline and then back online, it sends a subscription
report with Hue, Saturation, X, and Y values. The ColorMode
needs to be checked before emitting the colorControl capability
with the hue and saturation values. The issue is that the ColorMode
attribute is not guaranteed to be processed by the driver before
the other color attributes. Therefore, the ColorMode handler
should send a read request for the proper color attributes
after checking the ColorMode.