Make tests backwards compatible down to lua libs v10
Add backwards compatibility for matter switch test cases down to lua
libs version v10. Most of the issues were from test cases expecting
native handler registrations, which was added in v11. Also, running the
test cases against lower lua libs caught a bug in the driver code for
lua libs version v10 and lower, which prevented the energyMeter
capability from being emitted. This is also fixed by this PR.

It might be possible to extend support to even lower versions, but there
are a few additional api changes that would need to be accounted for.

Also note that to run these tests against lower versions, any
capabilities that are not available in that version of the lua libs
would need to be manually added to prevent test failures.