Improve Matter Thermostat component to endpoint mapping (#2099)
Add an optional argument to component_to_endpoint to specify a cluster
ID. The previous implementation checked for the presence of airPurifierFanMode
within this function, which fails to account for situations where a
device supports both the FanControl and Thermostat clusters on different
endpoints. Remove thermostatOperatingState from a profile that is not supposed to include it.
Implement handling for the resetFilter capability command.