Matter Switch move initialization to doConfigure (#2041)
This change moves the initialization logic for buttons and switches to
doConfigure. This keeps all of the profile selection logic all within
doConfigure and allows the removal of logic gates from device_init that
were there to ensure init code only ran one time.

Also added is a new function that runs at init that can rename or delete
persisted fields on the device.