Matter Thermostat: Use larger temperature range
For RPC >= 5, the hub converts temperature values from capability
commands to Celsius, meaning the driver no longer needs to use separate
ranges for Celsius and Fahrenheit in order to determine the unit as
Celsius can always be assumed. A range is still used, in order to filter
out values received from the device that likely to be erroneous, but it
is much larger and will be able to accommodate the temperature range for
most use cases.