Use static profile subdriver rather than modular subdriver
It might make more sense to put the code for environments that don't
support modular profiles into a subdriver so that it doesn't even need
to be loaded on setups that do support modular profiles. This way, the
main driver contains the most up to date code rather than having to maintain
two paths.