Add reconfiguration for zigbee power meter to reduce reporting
This will check devices on init to see if they support the clusters for power
reporting and if so schedules a 5 minute timer to check all devices in the
driver, and attempt to configure them to a lower reporting frequency.  It will
then mark the configuration version on the device to avoid it happening again.