fix: routing non-group commands to the correct speaker
I'm not sure if this is a Sonos API Change or if this bug has always
been here, but the driver was routing all outgoing websocket commands to
the group coordinator at all times.

This was causing non-group addressed commands, such as individual player
volume when in a group, to be sent over the wrong connection.

The net result was a no-op, because the JSON payload also includes a target
player ID, and when that ID doesn't match the player that received the
command, it gets ignored.