fix(sonos): Crashes that can occur on already pathological paths.
Release of the driver to general population has revealed that there are
various crashes along pathological code paths for the Sonos driver.
These are pathological paths that didn't show up in testing, and based
on where the fixes were made, these would be devices that are already
non-functional in some capacity.

The paths have been adjusted to avoid crashing where possible, with
increased logging, including some logging to hub-core where the nature
of the crash (typically unexpected input) is something that hasn't been
seen or reproduced yet, so that we can hopefully get more info in the
future.

These are the fixes that have been made:

- Bad `string.format` call on pathological log line

- Additional nil checking during group role inspection

- more robust handling around failed SSDP task spawning
  - Add a missing nil check
  - Send SSDP task failure logs to hub-core, since those are
    critical/irrecoverable.

- Input checking before extracing MAC from serial
  - Crashes are showing that we're getting discovery info that is lacking
    expected structure, and causing crashes.

- Log instead of crash on invalid key for `CaseInsensitiveKeyTable`s.

- Add checks around the security API for very old Lua Libs.

- Missed some locations where we should guard the oauth flow.