Off by 1 error prevents code scan complete event
This did not prevent any codes from not being scanned, but it did
prevent the scan complete event from being emitted. This also moves the
default handlers into the driver itself. We encountered one user who
didn't seem to have the defaults registered, and despite us not being
able to reproduce or find a bug, this seems like it fixes the issue for
that user.