Fix reverse preference in matter window covering
This change inverts the interpretation of open/closed when the Reverse
preference is set. This means that when the preference is set,
a `windowShadeLevel` of 100% is interpreted as Closed and 0% as Open. A
previous change that inverted the percentage values received from the
device is rolled back by this change because that did not properly
implement the reverse preference.