Show how to fix tests that expect device lifecycle events to occur
Note how we need only need to simulate the added, init, doConfigure, and
infoChanged flow because in those flows we "configure" the buttons by
setting various endpoint fields so we know how to handle various button
flows. If we were doing the field setting, we could leave the startup
messages enabled and use the mock object as is to validate the init
sends the correct subscribe and various cap cmd/protocol messages are
handled correctly.