Change main component name in keyboard profile (#2264)
This change uses a label to display the main component in the
12-button-keyboard profile as F1 while renaming the actual
component name to 'main'. This is needed because the first
component not being 'main' is causing an issue seen in the
plugin for this device. The issue appears to be limited to
Android and involves the plugin sometimes not rendering
the supported button values properly.