Reimplement reverse preference in matter window covering (#2301)
This change inverts the interpretation of open/closed when the
Reverse preference is set. This means that when the preference
is set, a `windowShadeLevel` of 100% is interpreted as Closed
and 0% as Open. This approach replaces a previous change
that inverted the percentage values received from the device,
which was more complex and has some edge cases that caused
occasional issues.