sonos: Fix websocket reconnect task
There is a race condition when a sonos connection calls close on itself.
It immediately calls on_close which will spawn a task to reconnect. The
reconnect task checks if the connection is running which depends on
the socket being closed and cleaned up. Instead, call on_close
after the socket has been closed and cleaned up.