Use fanSpeedPercent for thermostats
This change adds the fanSpeedPercent capability to the
thermostat-modular profile. Additionally, PercentSetting is subscribed
in addition to PercentCurrent to set this capability, since it provides an
accurate representation of the speed of the fan while helping avoid
the following situation:

(1) The fan speed is changed in the app and PercentSetting is routed to
the device
(2) The fan reports back a value of PercentCurrent that doesn't match
PercentSetting because the speed takes a little while to change
(3) The fanSpeedPercent capability jumps to the value reported by
PercentCurrent

PercentCurrent is still subscribed to, but its attribute handler is
gated on the current fan mode being AUTO, in which case PercentSetting
is NULL on the device side and PercentCurrent is used as a fallback for
setting the capability.